requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
overrides==7.7.0
prometheus_client==0.21.0
pyarr==5.2.0
pyinstaller==6.11.1
pytest==8.3.3
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
schema==0.7.7
setuptools==75.5.0
types-requests==2.32.0.20241016
urllib3==2.2.3
wheel==0.45.0